This is an alpha build.  I am working on adding Mono support.  This build has basic suport for OSX.  It uses Apple script to set the wallpaper after it has been downloaded.  It also adds support for tagging providers with OS tags so you can specify that a particular provider only works on a particular OS.
The only catch with this build on OSX is that the options screen doesn't work in OSX yet so if you want to change anything you are probably out of luck for the moment until I get things cleaned up.  That means on OSX you are stuck with Wallbase "Nature" wallpapers (anyone who can download and rebuild this alpha build should be able to work around those limitations though).
(So many changes to get this working in Mono, didn't want to lose them so checking them in)