I'm working to expand the underlying setup of input/output providers.  I want output providers to be able to request additional pictures if needed and not just work with a single one.  Output providers now handle requestingg however many pictures they need.
Also instead of a random provider being chosen for each picture request (when more then one is selected), now a random picture is chosen out of all active providers.

Part of the plan for allowing output providers to select their own pictures is that down the road when multi-monitor support is added you will be able to say "I want wallpapers from this provider on this screen", and the output providers will be able to do it.

I also added icons for all input providers.  These icons show up in the input provider drop down list and also in the list of active providers.